+++ src/web/assets/cp/src/form-builder/fields/StatusRuleConditionsField.jsx
@@ -1,0 +1,16 @@
+import { ConditionsFieldBase } from './ConditionsFieldBase';
+
+function StatusRuleConditionsField({ field, form }) {
+    return (
+        <ConditionsFieldBase
+            field={field}
+            form={form}
+            ruleKey="applyRule"
+            defaultRuleValue="apply"
+            hideRuleSelector
+            referenceContext="client"
+        />
+    );
+}
+
+export { StatusRuleConditionsField };