+++ src/workflow/stages/NormalizeStage.php
@@ -6,6 +6,7 @@ use verbb\formie\services\SubmissionWorkflow;
 use verbb\formie\workflow\StageInterface;
 use verbb\formie\workflow\StageResult;
 use verbb\formie\workflow\WorkflowContext;
+use verbb\formie\workflow\tasks\normalize\ApplyStatusRulesTask;
 use verbb\formie\workflow\tasks\normalize\ClearConditionallyHiddenFieldsTask;
 use verbb\formie\workflow\tasks\normalize\EnsureSubmissionDefaultsTask;
 use verbb\formie\workflow\tasks\normalize\HandleBackNavigationTask;
@@ -39,6 +40,7 @@ class NormalizeStage implements StageInterface
             new ResolvePageFlowTask(),
             new ClearConditionallyHiddenFieldsTask(),
             new EnsureSubmissionDefaultsTask(),
+            new ApplyStatusRulesTask(),
         ]);
     }
 }